+++ src/Sidebar/SidebarDelegate.swift
@@ -11,6 +11,9 @@ protocol SidebarDelegate: AnyObject {
     /// Called when user requests to eject a volume
     func sidebarDidRequestEject(_ volume: VolumeInfo)
 
+    /// Called when user requests to eject all volumes from a server
+    func sidebarDidRequestEjectServer(host: String)
+
     /// Called when user adds a folder to favorites via drag-drop
     func sidebarDidAddFavorite(_ url: URL, at index: Int?)